requirements.txt
selenium==3.12.0
PyYAML==3.12
